requirements.txt
anyio==3.6.2
APScheduler==3.10.0
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
